modelproject/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-debug-toolbar==6.0.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
packaging==25.0
pycparser==2.22
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
pymysql==1.1.1
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4

# 개발, 테스트, 배포까지 모두 공통으로 쓰는 패키지를 기록
# ex) Django, DRF, pytest, debug-toolbar 등